project/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
marshmallow==3.10.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
PyJWT==2.0.1